requirements.txt
streamlit
pandas
python-dotenv
supabase
matplotlib

